refactor(error-handler): reorganize components
